feat(inventory): export PDF workflow runtime contract
