+++ .changeset/inventory-workflow-runtime.md
@@ -1,0 +1,5 @@
+---
+"@voyant-travel/inventory": patch
+---
+
+Publish the product PDF workflow deployment runtime service contract for graph activation.
